refactor CreateCommand
